### 2022-02-02 #4124 @whedon
[PRE REVIEW]: smot: a python package and CLI tool for contextual phylogenetic subsampling
**Submitting author:** @arendsee (<a href="http://orcid.org/0000-0002-5833-798X">Zebulun Arendsee</a>)
**Repository:** <a href="https://github.com/flu-crew/smot" target ="_blank">https://github.com/flu-crew/smot</a>
**Version:** v0.17.3
