### 2022-02-02 #4124 @Kevin-Mattheus-Moerman
@openjournals/dev can you see what is wrong with the paper? Is it the use of `` `smot` `` in the title?
