### 2022-02-02 #4124 @danielskatz
@arendsee - you might want to try adding one or more tags, as in the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography)
