### 2022-02-02 #4124 @danielskatz
Use the command `@whedon check references` to check the references again, and the command `@whedon generate pdf` after making changes to the .md file or when the references are right to make a new PDF. Whedon commands need to be the first entry in a new comment.
