### 2022-02-02 ac3ddeb
Add DOIs to bibtex file
