+++ paper/paper.bib
@@ -6,7 +6,9 @@
   number={3},
   pages={e9490},
   year={2010},
-  publisher={Public Library of Science San Francisco, USA}
+  publisher={Public Library of Science San Francisco, USA},
+  doi={10.1371/journal.pone.0009490},
+  url={https://doi.org/10.1371/journal.pone.0009490}
 }
 
 @article{zhang2017influenza,
@@ -17,7 +19,9 @@
   number={D1},
   pages={D466--D474},
   year={2017},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/nar/gkw857},
+  url={https://doi.org/10.1093/nar/gkw857}
 }
 
 @article{anderson2016phylogeny,
@@ -28,7 +32,9 @@
   number={6},
   pages={e00275--16},
   year={2016},
-  publisher={Am Soc Microbiol}
+  publisher={Am Soc Microbiol},
+  doi={10.1128/mSphere.00275-16},
+  url={https://doi.org/10.1128/mSphere.00275-16}
 }
 
 @article{katoh2013mafft,
@@ -39,7 +45,9 @@
   number={4},
   pages={772--780},
   year={2013},
-  publisher={Society for Molecular Biology and Evolution}
+  publisher={Society for Molecular Biology and Evolution},
+  url={https://doi.org/10.1093/molbev/mst010},
+  doi={10.1093/molbev/mst010}
 }
 
 @article{vijaykrishna2010reassortment,
@@ -50,7 +58,9 @@
   number={5985},
   pages={1529--1529},
   year={2010},
-  publisher={American Association for the Advancement of Science}
+  publisher={American Association for the Advancement of Science},
+  url={https://doi.org/10.1126/science.1189132},
+  doi={10.1126/science.1189132}
 }
 
 @article{smith2009origins,
@@ -61,7 +71,8 @@
   number={7250},
   pages={1122--1125},
   year={2009},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1128/jvi.00459-15}
 }
 
 @article{nelson2015continual,
@@ -72,7 +83,8 @@
   number={12},
   pages={6218--6226},
   year={2015},
-  publisher={Am Soc Microbiol}
+  publisher={Am Soc Microbiol},
+  doi={10.1128/JVI.00459-15}
 }
 
 @article{volz2013viral,
@@ -83,7 +95,8 @@
   number={3},
   pages={e1002947},
   year={2013},
-  publisher={Public Library of Science San Francisco, USA}
+  publisher={Public Library of Science San Francisco, USA},
+  doi={10.1371/journal.pcbi.1002947}
 }
 
 @article{baum2005tree,
@@ -94,7 +107,8 @@
   number={5750},
   pages={979--980},
   year={2005},
-  publisher={American Association for the Advancement of Science}
+  publisher={American Association for the Advancement of Science},
+  doi={10.1126/science.1117727}
 }
 
 @article{baum2013tree,
@@ -102,7 +116,8 @@
   author={Baum, David A and Smith, Stacey D},
   journal={An Introduction to Phylogenetic Biology. Roberts and Company Publishers},
   year={2013},
-  publisher={Citeseer}
+  publisher={Citeseer},
+  doi={10.22269/210921}
 }
 
 @article{baldauf2003phylogeny,
@@ -113,7 +128,8 @@
   number={6},
   pages={345--351},
   year={2003},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/s0168-9525(03)00112-4}
 }
 
[…]
