### 2022-02-15 dbaf114
Fix install_requires entry in setup
