### 2022-02-16 #4124 @Bisaloo
Hi @marekborowiec :wave:, you signed up as a potential reviewer for JOSS. I think this submission could be relevant to your research interests:

> smot: a python package and CLI tool for contextual phylogenetic subsampling

Would you be available to review this [python package](https://github.com/flu-crew/smot) and [associated paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.04124/joss.04124/10.21105.joss.04124.pdf) for JOSS? :relaxed:
