### 2022-02-16 db2b95d
Do not import smot in setup for version (fix #1)
