### 2022-02-16 cd0ad8f
Make the linter happy with __version__
