### 2022-02-16 2eeb399
Typecheck after installation in CI workflow
