### 2022-02-16 c5b67ef
Install before unit tests also
