### 2022-02-16 575b54a
Replace the initial dummy value with a type annotation
