+++ paper/paper.bib
@@ -72,7 +72,7 @@
   pages={1122--1125},
   year={2009},
   publisher={Nature Publishing Group},
-  doi={10.1128/jvi.00459-15}
+  doi={10.1038/nature08182}
 }
 
 @article{nelson2015continual,
