### 2022-02-16 72492af
Only lint within smot/
