+++ .mypy.ini
@@ -1,0 +1,3 @@
+[mypy]
+
+files = smot/
