+++ .github/workflows/python-app.yml
@@ -27,10 +27,10 @@ jobs:
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
-        flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
+        flake8 smot/ --count --select=E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         # flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
-        flake8 . --count --exit-zero --max-line-length=127 --statistics --ignore=W503,E501,W605,E731,F401,E722
+        flake8 smot/ --count --exit-zero --max-line-length=127 --statistics --ignore=W503,E501,W605,E731,F401,E722
         # Installation must be done before unit testing or typechecking
         pip install .
     - name: Unit tests
