### 2022-02-23 #4124 @Bisaloo
Yes, that would be perfect, there is no huge rush and we are usually pretty flexible with deadlines.

I'll go ahead and add you as reviewer then. I'll open the actual review issue shortly after and you'll be able to read more detailed instructions about the review process at JOSS there. Thanks!
