### 2022-02-23 #4193 @editorialbot
[REVIEW]: smot: a python package and CLI tool for contextual phylogenetic subsampling
**Submitting author:** @arendsee (<a href="http://orcid.org/0000-0002-5833-798X">Zebulun Arendsee</a>)
**Repository:** https://github.com/flu-crew/smot
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
