### 2022-02-23 #4193 @Bisaloo
:wave: :wave: :wave: @arendsee @Chjulian @marekborowiec this is the review thread for the paper. All of our communications will happen here from now on.

We usually aim for reviews to be completed within about 2-4 weeks. Please let me know if any of you require some more time. We can also use editorialbot (our bot) to set automatic reminders if you know you'll be away for a known period of time.

If you have reviewed for JOSS in the past, please be aware that our workflow has slightly changed. You can now generate your review checklist by posting a comment with the following content:
