### 2022-03-02 #4193 @Chjulian
smot (Simple Manipulation Of Trees) helps colour, annotate, and subset a tree (or extract a subtree rather than sampling a tree). Still, it is unclear what the authors mean by summarising, in phylogenetics, often means to map multiple trees into a single one.

I think that smot could be helpful to reduce the complexity of a tree to help unveil and communicate evolutionary features.

Documentation

General comment:
Overall I think the -h and README.rd are not exhaustive or detailed enough. Given that this package deals with graphical tree-like structures, it requires more visual examples. The users should clearly understand what a command is intended for, as this is hard to tell from plain text. The provided examples are difficult to interpret because they are large phylogenies (Also, note that there are no details on what is happening in example 2 with each instruction). A smaller model, i.e. a tree with a dozen of taxa and the display of the most varied tree manipulations and key features (for example displaying the labels on the tree with each of the factoring schemes, displaying examples for each of the three ways of sampling, displaying the difference on using pull versus push when colouring, etc.) would give us an accurate idea of smot’s utility.

Specific comments:
README.md. I found it puzzling to list the subcommands rather than the commands with the corresponding subcommands/options. I also think that the distinction between a command, subcommand and option is not consistent across the help.

The installation worked for me, but post-installation, the package required two dependencies that were not flagged as required during the installation.
```
ModuleNotFoundError: No module named 'parsec'
ModuleNotFoundError: No module named 'click’
```

smot does not respond when I do not specify the `[TREE]`. Ideally, the user should receive an error or a warning.

The help informs that the command  ‘stat’ should ‘Display statistics for an input tree’, but it seems to return only the number of tip labels.

The command `grep -f` works when the grep expression is saved without quotation marks. I’m not sure if this is standard practice outside the command prompt, but it is worth checking.

Concerning the command sample:
I find it confusing that the options equal, mono and para are listed as commands. There is no further guidance on how to use them. `smot mono -h` returns the `Error: No such command 'mono'`.
Three factoring schemes appear in the descriptive text but are not listed as options, and I did not find documentation.
The examples of sample written to the README.md use the options `prop`, `min-tips`, `keep` and `seed`.  I did not find documentation of these options. Using `prop` triggers the `Error: No such command 'prop'`.

I think that some commands need to be chained in a particular order. This is not clear from the help or the README.md.

More details are needed for the command filter. For instance, it is not clear how the input should look like for the option `--factor-by-field`.

The help for `color` is not comprehensive. I had to look up the README.rd to work things out.
The description informs that the colouring options are highly opinionated. I’m unsure what this means, but I guess it means that there are no default palettes to sample colours from and that the user must indicate each colour one by one.
The options of `color` are listed as commands.
The example in the README.md uses `smot color leaf -P -p` but the options `-P` and `-p` are not documented in the help of `color`.
The help prints, `There is no simple way to colour a  particular branch (and why would you want to do that anyway). So, follow my tree coloring dogma and everything will be fine`. I kindly disagree as sometimes a user may need to colour a single tip because it is a reference, the target sequence, etc.

Software paper

General comments:
Some aspects of the methodology are not apparent because plain text description of tree-like structures is often not enough. Thus, some additional illustrations may be needed.

Statement of need. I am not sure all the info in the second paragraph fully achieves what it is intended to achieve (note that some of the points raised in this section are paraphrased (or missing) in the next section). Here, I understand that the authors must provide more details on the missing niche(s) that smot intends to fill. For instance, to inform if there is no software available to extract a subtree based on grep or regex and communicate why this could be beneficial (less time consuming, less error-prone, easy to include in near real-time pipelines analysis of sequence data, help to identify features from the tree, help communication of findings, etc.).

Specific comments:
I prefer not to use the term “(reference) strain” here because the meaning of strain changes across disciplines and taxonomic groups. I suggest using reference taxa/taxon as it is precise in phylogenetics.

Line 15. genetic (or molecular as the authors use elsewhere in the text) sequence dataset

Line 23. variables instead of parameters?

Line 25. It is not fully clear to me what the authors are implying here with classification because the authors mention, for instance, that geographical location may be incomplete and requires classification. By classification do you mean annotation? that is, the metadata is available but is not represented on the tree. Otherwise, I don’t see how you can make such attributions based on phylogenetic associations alone. One can attribute unknown taxa to a particular taxonomic group because we classify biological entities based on evolutionary relationships (and I think this is implied on lines 34 and 49).

Line 26. I’m unsure about correlation being the correct term here, as correlation is not the only type of relationship that phylogeneticists are looking for.  maybe something along the lines of “subsetting and colouring facilitates the communication and interpretation of phylogenies and traits.”

Line 33 r is not defined. Not clear how the scaling works here. Do you mean to reduce the size of each monophyletic group by retaining (if possible) a user-defined number of tips?

Line 35. I am not sure about the use of ‘ingest’. Maybe ‘take’ or ‘process’?

Line 36. I think the authors mean, “remove any monophyletic groups that do not meet certain criteria (lack or a variable value) or size”.

Line 50. What do you mean by purity? Groups are either monophyletic or not.

Line 68. Core algorithm. The paraphyletic versus Monophyletic algorithm is not fully clear to me. I think illustrations would be of great help here. This is important to understand line 127 too.

Plots in figure 1 are mislabelled. I think D corresponds to (missing) C and E corresponds to D.

Line 53. It is unclear what the authors explain here and what they call a hybrid approach. I understand that, given a phylogeny (built with reference taxa), smot attributes the unlabeled taxa to any taxonomic reference groups.

Line 56. Details are not provided about filtering. Also not clear what exactly smot does here concerning colouring. I understand that it creates a nexus output that can be read using FigTree?

Line 126. What is the meaning of n?

In line 119, by “all swine clades with a single member”, do you mean singletons?

Line 125. To understand these more complex trees and the implications of the methods, I need more info on the paraphyletic versus Monophyletic algorithm (schematics).
