### 2022-03-13 95a7d1d
Mention subcommand help in README
