### 2022-03-13 b3d89c9
Add painfully obvious subcommand help statements
