### 2022-03-13 de57d31
Remove stubby stat command
