+++ smot/main.py
@@ -187,22 +187,6 @@ def tips(tree: TextIO) -> None:
         print(tip)
 
 
-@click.command()
-@dec_tree
-def stat(tree: TextIO) -> None:
-    """
-    Display statistics for an input tree
-    """
-    import smot.algorithm as alg
-
-    tree_obj = read_tree(tree)
-    tree_obj.tree = alg.setNLeafs(tree_obj.tree)
-
-    tips = list(alg.tips(tree_obj.tree))
-
-    print(len(tips))
-
-
 def factoring(function):
     function = click.option(
         "--factor-by-capture",
@@ -1147,7 +1131,6 @@ color.add_command(push_color)
 color.add_command(pull_color)
 
 cli.add_command(tips)
-cli.add_command(stat)
 cli.add_command(sample)
 cli.add_command(factor)
 cli.add_command(tipsed)
