+++ smot/version.py
@@ -1,1 +1,1 @@
-__version__ : str = "0.17.3"
+__version__: str = "0.17.3"
