### 2022-03-21 #4193 @Bisaloo
Hi @marekborowiec :wave:, friendly reminder that your feedback on this [piece of software](https://github.com/flu-crew/smot) and [related paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.04193/joss.04193/10.21105.joss.04193.pdf) would be appreciated.

I understand if you are busy so feel free to let me know if you need more time. This message is mainly to keep this on your radar :relaxed:
