### 2022-04-06 #4193 @Bisaloo
Hi @marekborowiec :wave:, are you still available to review this [piece of software](https://github.com/flu-crew/smot) and [related paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.04193/joss.04193/10.21105.joss.04193.pdf)?

Please let me know if you need more time or if you don't think you will be able to make it :relaxed:
