### 2022-04-07 #4193 @marekborowiec
I was not able to install using `pip`. The `README` file states the requirements are `biopython`, `parsec`, and `docopt`. The `requirements.txt` has `parsec` and `click` but these are not installed when using `pip`. I was able to at least run `smot` in a clean environment after installing `biopython`, `parsec`, and `click` first. In addition to the requirements being consistent and installed along with `smot`, the documentation could link to each dependency repository. In short, installation requirements need to be improved.

`smot` has the potential to be a fast and useful tool but I agree with @Chjulian that documentation needs to be better thought-out and improved before this software can be used by others. It needs to state somewhere which tree file formats are supported. Hard to get started without that. I work with phylogenetic trees every day and yet I am having difficulty understanding much of this program's functionality. Examples isolated to individual commands with smaller trees would go a long way.

Going over the individual commands, starting with `color` . using `smot color -h` I lean that "coloring options are highly opinionated" and "follow my tree coloring dogma and everything will be fine". Alright, but there is no explanation of how the coloring algorithm works, leaving the user in the dark. I tried to run the command with `smot color leaf mytree.nexus > colored.tre`. According to the documentation, this should have produced a tree with colored tips. I see no modification to the tree when printed to the screen. When redirected and opened using `FigTree` I also did not see any coloring. Using `smot color branch mytree.nexus > colored.tre` did not work at all and it wasn't until I ran `smot color branch -h` I learned that I may provide a color map (but there are no instructions on how to do this) and that if I don't, `smot` will automatically apply a color-blind friendly palette. As far as I can tell, however, no modification is done to my tree using `smot color branch mono mytree.nexus`.

The commands `factor` and `sample` seem to be important and potentially very useful but I cannot understand how they work after looking at the examples provided and reading through in-program documentation. Small examples of both trees and table inputs will be helpful in explaining here.

I like the `grep`, `tips`, and `tipsed` commands, which are simple and seem to work as advertised. Very useful.

`stat` seems to only provide the number of tips and it is not clear that this is what it does until one runs it. I would have expected things like tree length, average branch lengths, etc.

As for the paper, I would say provide an exhaustive list of program capabilities and compare it with other available software. You say that `smot` "complements the existing Python phylogenetics ecosystem through its subsampling, classifying, and filtering algorithms" but what does that mean, exactly? Are DendroPy or TreeSwift not able to do any of these actions? What can they do instead?

I think these are the main issues I can identify right now. I hope the authors find my comments useful and I commend them to submitting to JOSS with the intent of improving this promising piece of software.
