### 2022-04-24 #4193 @arendsee
I would like to thank @Chjulian and @marekborowiec  for their thoughtful comments. Sorry for the delay, I've been transitioning into a new position and a new city, so am a bit back-logged.

Below are my responses to the main points the raised with respect to
the `smot` code. I will address the comments on the paper next. All the changes
I mention below are in smot v0.17.4.

## It is not clear what the author meant by "summarizing" the trees

Good point, I have removed the phrase from the README and program description.
When I wrote that description, I had more ambitious plans for the `smot stat`
function for returning tree statistics and summaries.

## The `smof stat` function just prints the number of tips

Right, that command was a stub. As mentioned above, I had intended to make it
more sophisticated. Instead I will just remove it. If you want to know the
number of tips in a tree, you can just call `smot tips x.tre | wc -l`.

## Some options or commands are not documented

The reviewers misunderstood the nested documentation paradigm of smot. This is
a problem, of course, because if they misunderstand others will as well. For
this reason, I have added a help statement at each internal documentation level
that gives an example of how to access subcommand help. For example:

```
$ smot color -h
Usage: smot color [OPTIONS] COMMAND [ARGS]...

  Color the tips or branches.

  The coloring philosophy in smot is to encode information about the specific
  taxon in the taxon label color and to encode data about groups of taxa in
  the branch color.

  If you want to color tips by phylogenetic group, you may first color by
  branch and then push the colors to the tips:

    smot color branch mono --factor-by-field 1 -c colormap foo.tre |     smot
    push

  This will identify all monophyletic groups based on the clade name found in
  the first position of the taxon label (e.g., `>primate|...`. The
  monophyletic branches are colored based a color map, which should have an
  entry such as `primate #00FF00`. Then `smot push` takes colors in the nodes
  and pushes them to their children, thus pushing the parent node colors to
  the taxon labels.

  Conversely, taxon label colors can be pulled up into parent nodes using
  `smot color pull`.

Options:
  -h, --help  Show this message and exit.

Commands:
  branch  Color the branches of a tree.
  leaf    Color the taxa labels on a tree.
  pull    Pull colors from tips to nodes
  push    Push colors from nodes to tips
  rm      Remove all color annotations from a tree

  For subcommand usage, append `-h` (e.g., smot color branch -h)
```

The last line of the help statement gives can example of how to access
subcommand help. Following that line:

```
$ smot color branch -h
Usage: smot color branch [OPTIONS] COMMAND [ARGS]...

  Color the branches of a tree. You may provide a color map; if you do not,
  smot will automatically map factors to colors from a color-blind friendly
  palette.

Options:
  -h, --help  Show this message and exit.

Commands:
  mono  Color a tree by monophyletic factor.
  para  Color a tree by paraphyletic factor.

  For subcommand usage, append `-h` (e.g., smot color branch mono -h)
```

This takes us to a terminal command:

```
$ smot color branch mono -h
Usage: smot color branch mono [OPTIONS] [TREE]

  Color a tree by monophyletic factor.

  smot color branch mono --factor-by-capture="(1B\.[^|]*)" 1B.tre

Options:
  --factor-by-table ?STR    Read factors from a TAB-delimited file (tip labels
                            in the first column and factors in the second
                            column)
  --factor-by-field ?NAT    Factor by 1-based field index (with '|'
                            delimiters, for now)
  --factor-by-capture ?STR  A regular expression with a capture for
                            determining factors from labels
  -c, --colormap PATH       A TAB-delimited, headless table with columns for
                            factor names and hexadecimal colors
  -h, --help                Show this message and exit.
```

Each level of the documentation shares what is common to all subcommands. At
the terminal level of documentation, the options are listed. This approach to
documentation avoids the long dump of often conflicing arguments that is common
in command line tools (e.g., blast).

## The documentation is not extensive enough (especially for the color command)

Very true, I've gone through all the internal documentation in the package and
clarified it. The color documentation I have completely rewritten. I have added examples to most of the commands.

## Pip installation required manually installing parsec and click

Fixed in v0.17.4

## The command grep -f works when the grep expression is saved without quotation marks.

    It is standard practice. Specifically, UNIX grep, which `smot grep` is
    modeled on, works without quotes. Quotes are only necessary when the
    pattern contains spaces or characters that have meaning in the shell (such
    as '>' or '\').

## The program hangs when no file is given

This is because it is waiting for input from STDIN. UNIX grep and most other tools that can read from STDIN do the same. If you really wanted to, you could type a tree into terminal. Just for fun, run `smot tips`, type `(A,B);` and then hit Ctrl-D.

I think that hits the points raised about the program itself. If you have any additional comments, I'm happy to address them. I'll get to the paper as soon as I can.
