### 2022-04-24 843b512
Add examples and clearer explanations across help
