+++ smot/main.py
@@ -311,6 +311,10 @@ def equal(
     the depth of the tree, but is not representative of the tree's breadth.
     That is, if N=6 and a tree splits into two subtrees, one with 3 tips and
     one with 300 tips, still 3 tips will be sampled from each branch.
+
+    Example:
+
+      smot sample equal --factor-by-capture="(1B[^|]*)" --max-tips=5 1B.tre
     """
 
     import smot.algorithm as alg
@@ -361,9 +365,31 @@ def mono(
     tree: TextIO,
 ) -> None:
     """
-    Proportional sampling. Randomly sample p (0 to 1, from --proportion) tips
-    from each monophyletic (relative to factors) subtree. Retain at least N
-    tips in each branch (--min-tips).
+    Monophyletic sampling. Randomly sample --proportion of the tips (0 to 1)
+    from each monophyletic subtree. Retain at least --min-tips tips in each
+    branch.
+
+    The branches that are subsampled are monophyletic with respect to a string
+    (such as a clade name) that is identified by --factor-by-field,
+    --factor-by-table, or --factor-by-capture.
+
+    Examples:
+      \b
+      # downsample to the max of 10% or 3 taxa from each clade in a 1B influenza tree 
+      smot sample mono --factor-by-capture="(1B\.[^|]*)" -p 0.1 --min-tips=3 1B.tre
+
+      \b
+      # downsample each branch to the cube root (good when some clades are extremely large).
+      smot sample mono --factor-by-capture="(1B\.[^|]*)" -s 3 --min-tips=3 1B.tre
+
+      \b
+      # downsample each branch to a random 3 members or fewer members
+      smot sample mono --factor-by-capture="(1B\.[^|]*)" -s 3 1B.tre
+
+      \b
+      # downsample 1B tree but keep taxa with the pattern "202.-..-.." (taxa
+      # from the 20s) in their labels or the factor "1B.1.1". 
+      smot sample mono --factor-by-capture="(1B\.[^|]*)" -p 0.1 -r "202.-..-" -k "1B.1.2" --min-tips=3 1B.tre
     """
 
     import smot.algorithm as alg
@@ -433,6 +459,20 @@ def para(
     different factor than the stored subtrees, then we stop and sample from all
     tips in the stored trees and initialize a new list with the new
     monophyletic child.
+
+    Example:
+      \b
+      # Sample 10% of the members in each paraphyletic human or swine grouping
+      smot sample para --factor-by-capture="(swine|human)" -p 0.1 --min-tips=5 pdm.tre
+
+      \b
+      # Sample the square root of the number of members in each paraphyletic
+      # human or swine grouping
+      smot sample para --factor-by-capture="(swine|human)" -s 2 --min-tips=5 pdm.tre
+
+      \b
+      # Sample 3 of the members in each paraphyletic human or swine grouping
+      smot sample para --factor-by-capture="(swine|human)" -n 3 pdm.tre
     """
 
     import smot.algorithm as alg
@@ -492,7 +532,11 @@ def factor(
     example, for inferring clades given a few references in a tree. There are
     three modes: 'table' prints a TAB-delimited table of tip names and factors,
     'prepend' adds the factor to the beginning of the tiplabel (delimited with
-    '|'), 'append' adds it to the end.
+    '|'), and 'append' adds it to the end.
+
+    For example, to create a table of 1B clades: 
+
+      smot factor table --factor-by-capture "(1B[^|]*)" 1B.tre
     """
 
     import smot.algorithm as alg
@@ -712,7 +756,21 @@ def filter_cmd(
     """
     An advanced tool for performing actions (remove, color, sample, or
     replace) on monophyletic groups that meet specified conditions (all-match,
-    some-match, etc).
[…]
