### 2022-04-25 #4193 @Bisaloo
> The reviewers misunderstood the nested documentation paradigm of smot. This is
a problem, of course, because if they misunderstand others will as well.

:+1:

This might also be the sign that you should use another (additional) framework for the documentation. It's nowadays pretty common to have a documentation website (typically readthedocs for python packages but I'm sure they are plenty of alternatives) alongside the bundled documentation. A website might make it easier to navigate a nested/hierarchical structure. What do you think?
