### 2022-04-26 #4193 @arendsee
@Bisaloo Online documentation is nice for case studies and high-level descriptions where diagrams and figures are helpful, but I would always prefer quick CLI usage statements if I need to know what flags are available. The smot CLI documentation is easy and minimal; I just hadn't explained it.
