### 2022-05-02 f350af3
Fix fig1 and address reviewer comments in paper
TODO
 [ ] Re-read the paper, fix any unclear sentences
 [ ] Make a schematic of the equal, mono, and para algorithms
 [ ] Integrate schematic into the paper
