+++ paper/paper.md
@@ -27,7 +27,7 @@ bibliography: paper.bib
 `smot` (Simple Manipulation Of Trees) is a command line tool and Python package
 with the pragmatic goal of distilling large-scale phylogenetic data to
 facilitate visualization without jeopardizing inference. This package offers
-subsampling algorithms that preserve reference strains and tree topology,
+subsampling algorithms that preserve reference taxa and tree topology,
 algorithms for classifying unlabeled tips given a subset of labeled reference
 tips, and functions for filtering phylogenetic trees. The `smot` tool has broad
 application in phylogenetic analysis and we demonstrate its utility using a
@@ -45,8 +45,8 @@ example, in a phylogenetic tree with thousands of taxa that are clustered into
 many monophyletic groups, the user may want to subsample the taxa while
 ensuring all groups and their evolutionary relationships are represented in the
 visualized tree. Alternatively, taxa on trees may be described and grouped by
-parameters not defined by common ancestry such as geographical regions,
-phenotypes, sequence motifs, or a host species and this information may be
+variables not defined by common ancestry such as geographical regions,
+phenotypes, sequence motifs, or host species and this information may be
 incomplete and require classification. In these cases, subsampling and
 classifying data on the phylogenetic tree can form the basis of hypotheses on
 how temporal, spatial, and other processes correlate with the evolutionary
@@ -54,17 +54,18 @@ history of the studied population [@baum2005tree; @baum2013tree]. Developing
 appropriate hypotheses can be facilitated through the subsampling, classifying,
 and filtering algorithms in `smot`.
 
-The utility of `smot` is explained by the following three cases. For subsampling,
-`smot` may process a phylogenetic tree that has labeled clades: taxa from each
-clade may be subsampled while maintaining a set number of taxa from each clade
-and also retaining provided reference strains. Alternatively, the number of
-strains within a clade may be scaled by sampling the $r^{th}$ root. For
-classification, `smot` can process a partially labeled tree by inferring missing
-labels and prepending them to the taxa names. For filtering, `smot` may ingest a
-tree with user-provided trait labels and then remove any monophyletic group
-that has no recent member or has too few members. The subsequent section will
-briefly introduce related tools currently applied in phylogenetic analyses and
-outline the role `smot` can play in a phylogeneticist’s toolbox.
+The utility of `smot` is explained by the following three cases. For
+subsampling, `smot` may process a phylogenetic tree that has labeled clades:
+taxa from each clade may be subsampled while maintaining a set number of taxa
+from each clade and also retaining provided reference taxa. Alternatively, the
+number of strains within a clade may be scaled as $n_{sampled}=n^{1/r}$, where
+$n$ is the original number of taxa in the clade and $r$ is a user provided scaling
+factor.  For classification, `smot` can process a partially labeled tree by
+inferring missing labels and prepending them to the taxa names. For filtering,
+`smot` takes a tree with user-provided trait labels and then removes, subsamples
+or edits any monophyletic subtrees that meet specific criteria. The subsequent
+section will briefly introduce related tools currently applied in phylogenetic
+analyses and outline the role `smot` can play in a phylogeneticist’s toolbox.
 
 
 
@@ -83,14 +84,14 @@ tree topology.
 
 The second purpose of `smot` is classification and annotation of unlabeled taxa
 and clades when provided representative strain classifications. Classification
-is achieved by either patristic distance or monophyletic purity, and the method
-is based on submitted reference taxa rather than inferring clusters from the
-tree and then naming them. Inferring clusters de novo can be accomplished with
[…]
