### 2022-05-09 #4193 @Bisaloo
I'll let the reviewers weigh in but I'm not sure explaining it will be enough to make it more easily navigable. It'll always be difficult to have a general overview of all the possible options, without going inside each one of the items.

Note that I'm not arguing you should remove the CLI documentation. I'm recommending you add an alternative, which some users may find easier to understand.
