### 2022-05-29 c2dd278
Clean up documentation and data
