+++ smot/__init__.py
@@ -17,7 +17,7 @@ from smot.algorithm import (
     sampleRandom,
     sampleMonophyletic,
     sampleParaphyletic,
-    sampleBalanced,
+    sampleEqual,
     colorTree,
     colorMono,
     colorPara,
@@ -53,7 +53,7 @@ __all__ = [
     "sampleRandom",
     "sampleMonophyletic",
     "sampleParaphyletic",
-    "sampleBalanced",
+    "sampleEqual",
     "colorTree",
     "colorMono",
     "colorPara",
