+++ smot/main.py
@@ -281,7 +281,7 @@ dec_patristic = click.option(
 )
 
 
-@click.command()
+@click.command(name="equal")
 @factoring
 @dec_keep
 @dec_default
@@ -289,7 +289,7 @@ dec_patristic = click.option(
 @click.option("--zero", is_flag=True, help="Set branches without lengths to 0")
 @dec_newick
 @dec_tree
-def equal(
+def sample_equal_cmd(
     factor_by_capture: Optional[str],
     factor_by_field: Optional[int],
     factor_by_table: Optional[str],
@@ -301,16 +301,17 @@ def equal(
     tree: TextIO,
 ) -> None:
     """
-    Equal sampling. Descend from root to tip. At each node, determine if each
-    subtree contains a single factor. If a subtree is not monophyletic, recurse
-    into the subtree. If the subtree is monophyletic, then select up to N tips
-    (from the --max-tips argument) from the subtree. The selection of tips is
-    deterministic but dependent on the ordering of leaves. To sample a subtree,
-    an equal number of tips is sampled from each descendent subtree, and so on
-    recursively down to the tips. The resulting downsampled subtree captures
-    the depth of the tree, but is not representative of the tree's breadth.
-    That is, if N=6 and a tree splits into two subtrees, one with 3 tips and
-    one with 300 tips, still 3 tips will be sampled from each branch.
+
+    Equal sampling. Descend from root to tip. At each node, determine if all
+    descends have the same factor (is monophyletic). If the tree is not
+    monophyletic, recurse into each subtree. If the tree is monophyletic, then
+    sample N taxa from the tree.
+
+    To sample N taxa from a tree with k immediate subtrees, divide N as evenly
+    as possible into k groups. If a given subtree has fewer taxa than are
+    alloted to it, then keep all taxa in the subtree and divide the remainder
+    among its siblings. This partitioning is dependent on the ordering of the
+    tree, so for reproducibility the tree should be sorted before sampling.
 
     Example:
 
@@ -327,7 +328,7 @@ def equal(
         factor_by_table=factor_by_table,
         default=default,
     )
-    tree_obj.tree = alg.sampleBalanced(tree_obj.tree, keep=keep, maxTips=max_tips)
+    tree_obj.tree = alg.sampleEqual(tree_obj.tree, keep=keep, maxTips=max_tips)
 
     if newick:
         print(sf.newick(tree_obj))
@@ -335,7 +336,7 @@ def equal(
         print(sf.nexus(tree_obj))
 
 
-@click.command()
+@click.command(name="mono")
 @factoring
 @dec_keep
 @dec_keep_regex
@@ -348,7 +349,7 @@ def equal(
 @dec_newick
 @click.option("--zero", is_flag=True, help="Set branches without lengths to 0")
 @dec_tree
-def mono(
+def sample_mono_cmd(
     factor_by_capture: Optional[str],
     factor_by_field: Optional[int],
     factor_by_table: Optional[str],
@@ -369,9 +370,12 @@ def mono(
     from each monophyletic subtree. Retain at least --min-tips tips in each
     branch.
 
-    The branches that are subsampled are monophyletic with respect to a string
-    (such as a clade name) that is identified by --factor-by-field,
-    --factor-by-table, or --factor-by-capture.
+    Monophyletic subtrees are identified by walking from root to tip and
+    checking at each node whether all descendents have the same label (or no
+    label).
+
+    There is no requirement that each monophyletic subtree have a distinct
+    label. Therefore, polyphyletic cases are also handled by this algorithm.
 
     Examples:
       \b
@@ -422,7 +426,7 @@ def mono(
         print(sf.nexus(tree_obj))
 
 
-@click.command()
+@click.command(name="para")
 @factoring
[…]
