### 2022-07-06 #4193 @marekborowiec
Hi @Bisaloo.  There are still issues with installation. The README says it needs Python v3.6 or newer but I tried with a 3.6.2 installation and also a clean conda env with 3.6.15 and no luck, there is an issue with `from __future__ import annotations`. This statement won't work unless you have Python >3.7 and so the documentation is confusing on that point.  Also, the documentation states requirements are Biopython, parsec, and docopt but installing with pip in a clean environment installs parsec, zipp, typing-extensions, importlib-metadata, click, smot. I haven't tried all the functionality of the new version, but will it be compromised without Biopython?

Again, there needs to be a discussion of what tree formats smot can take. Several of the individual, nested help messages now say that they will only work with "nexus", for example, but that should be discussed in README/manuscript.

The nested nature of the documentation was pretty clear once I played around with the program. However, the CLI help should state how to us it up front. It seems this is improved now with added statements about nested help messages. My main criticism wasn't related to the layout of command help, just the fact that it is very difficult to understand how to use the many commands and subcommands and options without examples. The documentation is now improved now and it explains more with examples. At the same time, it would still be good to have a birds-eye overview of functionalities of the program so that potential users can see what it can do. As I already mentioned, the examples in README and manuscript are too complex and they obscure the functionality of the program. This could be done in a much more expanded README or, as @Bisaloo suggested, readthedocs or a similar platform. Providing something with smaller examples of each function would increase the program's visibility and user base, so it is in your best interest to work this out. You cannot expect every potential user to first install your program without having a good idea of what it can do, then go through the massive tree of help messages to figure out what they can do with it. No one does that, and so you can expect fewer users/citations for your program.

Issues in the manuscript have not been addressed yet.

Small point, throughout the documentation "leafs" should be "leaves".
