### 2022-07-18 878267f
Update installation instructions in README
