### 2022-07-30 f3e8e19
Addressed almost all comments in the paper
The last thing to do is add the diagram
