+++ paper/paper.md
@@ -128,9 +128,11 @@ algorithm also descends from root to tip, but rather than setting a
 monophyletic subtree to a group, it merges adjacent monophyletic groups with
 the same label down the tree. When a node is reached that is monophyletic for
 the two subtrees with different labels, each subtree is set as a group,
-ensuring that the branch nearest to a group border is sampled from.
+ensuring that the branch nearest to a group border is sampled from. See **Figure
+1** for an simple example of the difference between the monophyletic and
+paraphyletic grouping algorithms.
 
-![Interspecies transmission and evolution of the 2009 H1N1 influenza A virus pandemic (H1N1pdm09) lineage in swine and humans. (A) An inferred phylogenetic tree of influenza A virus (IAV) in swine hemagglutinin (HA) genes from the H1N1pdm09 lineage collected between 2015 and 2021. There are too many swine strains in the tree to read the labels even omitting the human influenza A virus H1N1 HA sequences necessary to capture the correct evolutionary context of the lineage. (B) An inferred phylogenetic tree of H1N1pdm09 lineage HA genes from humans and swine (26,802 genes, human in black, swine in orange). The tree is too large to see individual labels, and critical human-to-swine evolutionary linkage is obscured. To identify the evolutionary history of this IAV lineage, we include all swine HA genes to demonstrate onward transmission of the virus, and human HA genes to detect directionality of interspecies transmission. (C) An application of `smot`: human HA genes were down-sampled while keeping all swine genes. This ensured the context of human HA genes, allowing identification of human-to-swine spillovers and visualization of swine-to-swine transmission of the H1N1pdm09 lineage. All swine clades present in (B) are present in (C). (D) Using this approach, we identified a human-to-swine event (arrow) that seeded onward transmission in swine, followed by a single human HA gene nested within a monophyletic swine group (triangle) (blue rectangle in (C) and enlarged in the inset (D)). The human HA gene demonstrates a zoonotic (swine-to-human) transmission event. Subsampling human HA genes before building the tree or without considering context would likely obscure these two-way interspecies transmission events.\label{fig:f1}](img/fig1.png)
+![The monophyletic and paraphyletic algorithms differ in how they group the clades that will be downsampled (or otherwise acted upon). The paraphyletic algorithm groups adjacent monophyletic subtrees down the trunk but preserves the deepest monophyletic tree to guarantee that the nearest relative to a change in label is preserved.\label{fig:fig1}](img/fig1.png)
 
 Once a tree is partitioned into groups, it may be subsampled, classified, or
 filtered. Subsampling takes each partition and randomly selects either a set
@@ -143,6 +145,7 @@ labeled tip using a patristic classifier. Filtering performs an operation on
 each group under some condition, for example, it may delete all groups that
 have fewer than $n$ members.
 
[…]
