### 2022-10-05 flu-crew/smot#4 @Bisaloo
Provide an overview of the package functionality
This could take several forms:

- I have suggested having a non-nested documentation in the form of a readthedocs website

- @marekborowiec has suggested adding more examples to the README

Feel free to take the approach you prefer but there needs to be some form of overview of the package functionality to convince users to install & try your program.

This is part of https://github.com/openjournals/joss-reviews/issues/4193.
