### 2022-10-05 flu-crew/smot#5 @Bisaloo
Test examples of part of the continuous integration
I played around a bit more with the program and found an example that doesn't run.

> smot color branch mono --factor-by-field 1 -c colormap foo.tre | smot push

in `smot color -h`

There might be more errors in other places.

It would be good to add the examples to your continuous integration set up to ensure the documentation is always providing correct examples.

This is part of https://github.com/openjournals/joss-reviews/issues/4193.
