### 2022-10-05 #4193 @Bisaloo
Hi @arendsee, I have opened two issues summarising what seem to be the current blockers to get things moving.

Once this is solved, I can ask the reviewers to have another look at their checklists and see what's left to get this submission thought the door.

📝 [Checklist for Chjulian](https://github.com/openjournals/joss-reviews/issues/4193#issuecomment-1048945444)
📝 [Checklist for marekborowiec](https://github.com/openjournals/joss-reviews/issues/4193#issuecomment-1092189926)

I understand you are particularly busy in this moment but could you provide us with a tentative timeline for these changes please?
