### 2022-10-06 #4193 @arendsee
Thanks @Bisaloo, I saw your issues and can address them next week. I guess I can go ahead and write up a dedicated docs page. It isn't too much work. I probably just flatten the internal documentation and include output images of the trees.
