### 2022-10-11 arendsee/jon#5 @arendsee
The examples are correct, you just need to substitute real files. I assume the error message that was raised said that the `colormap` or `foo.tre` files did not exist?
