### 2022-10-11 arendsee/jon#5 @arendsee
I just added automatic tests of the examples. And actually you were right about that example being incorrect, it needs to be `smot color push`.
