### 2022-10-12 f2383cd
Add example test (Fix #5)
