+++ test-data/test-examples.sh
@@ -1,0 +1,9 @@
+#!/usr/bin/env bash
+
+set -e
+
+while read -r example
+do
+    echo $example > /dev/stderr
+    eval $example > /dev/null
+done < <(grep "^      smot" ../smot/main.py)
