+++ README.md
@@ -120,3 +120,170 @@ smot sample prop --proportion=0.1 --min-tips=2 --factor-by-capture="(human|swine
 # image D
 smot sample para --proportion=0.1 --min-tips=2 --factor-by-capture="(human|swine)" --keep="swine" --seed=42 pdm.tre > pdm-para.tre
 ```
+
+## Examples by subcommand
+
+This section will show examples for each subcommands. All of the examples can
+be executed using sample files available in the `test-data` folder in this
+repo.
+
+### `smot grep` - prune a tree to preserve only the taxa matching a pattern
+
+Create a subtree where all taxa are from the year 2020.The '-..-..' pattern is
+a regular expression that matches the MONTH-DAY syntax in a date and prevents
+most unwanted matches.
+
+```
+smot grep --perl '2020-..-..' 1B.tre
+```
+
+### `smot factor` - Impute, annotate with, and/or tabulate factors
+
+Associating "factors" with taxa is used across `smot` for annotating, sampling
+and coloring. These factors can be specified in one of several ways.
+
+ * `--factor-by-capture` uses a regular expression capture group to set the factor
+ * `--factor-by-field` uses the index of a delimited field in the taxon label
+ * `--factor-by-table` uses an two-column, TAB-delimited table that maps taxon name to factor
+
+The `smot factor` subcommand can list the extracted factors in a table:
+
+```
+$ smot factor table --factor-by-capture "(1B[^|]*)" 1B.tre | head
+publicIAV|A/swine/Minnesota/A02245706/2020|H1N2|swine|USA|1B.2.1|2020-07-30     1B.2.1
+publicIAV|A/swine/Minnesota/A02245749/2020|H1N2|swine|USA|1B.2.1|2020-10-15     1B.2.1
+publicIAV|A/swine/Minnesota/A02245719/2020|H1N2|swine|USA|1B.2.1|2020-09-28     1B.2.1
+publicIAV|A/swine/Minnesota/A02524813/2020|H1N2|swine|USA|1B.2.1|2020-10-27     1B.2.1
+```
+
+Or it can return the input tree with the factors appended or prepended to the
+taxon name (`smot factor append` and `smot factor prepend`, respectively).
+
+### `smot sample` - subsample the tree
+
+`smot` offers three general sampling algorithms.
+
+ * The `equal` sampling method samples an equal number of taxa from each branch
+   descending from each node if sufficient taxa are available in each branch.
+
+ * The monophyletic sampling method (`mono`) subsamples a branch of the tree if
+   all taxa within the branch have the same factor.
+
+ * The paraphyletic sampling method (`para`) allows adjacent branches that are
+   monophyletic to be sampled together. The algorithm guarantees that at least
+   one taxa is sampled from the deepest branch neighboring a different group.
+
+
+The following three commands all do the same thing (25% sampling from each of the 1B flu clades) using different factoring methods:
+
+```
+smot sample mono --factor-by-field=6 -p 0.25 1B.tre
+smot sample mono --factor-by-table 1B.tab -p 0.25 1B.tre
+smot sample mono --factor-by-capture "\|(1B\.[^|]*)" -p 0.25 1B.tre
+```
+
+For reproducibility, you can specify the random seed (e.g., `--seed=24601`).
+
+It may be useful to prevent certain taxa from being downsampled. For example
+reference taxa or taxa that are from a particular time or region of interest.
+This can be achieved with the `-k/--keep` and `--r/--keep-regex` options which
+prevents downsampling of specific factors or taxa with labels matching a given
+regular expression. In the example below, a tree is downsampled by clade
+(`1B.*`) but all taxa from the 20s and all taxa from the specific clade
[…]
