### 2022-10-22 #4193 @arendsee
@Bisaloo I went with adding more examples to the README instead of building a whole readthedocs site. The new examples give an overview of what can be done with `smot`. All the data needed to run the examples is present in the `test-data` folder, so anyone interested can explore the results using their choice of visualization software.
