+++ README.md
@@ -99,7 +99,15 @@ smot sample para select-swine.tre --scale=4 --factor-by-capture="(swine|human)" 
   --min-tips=3 --keep="swine" --seed=42 > select-swine-para-sample.tre
 ```
 
-Line 1 filters out all leaves with hosts other than swine or human. Line 2 removes all monophyletic swine clades that have no representative from 2021. Line 3 removes all swine clades represented by a single member. Line 4 removes any coloring in the input tree. Lines 5 colors all remaining swine clades orange (hexadecimal code “#FFA000”). Line 6 colors leaves gray by default, then colors swine orange and finally recent swine blue. Lines 8-9, 10-11, and 14-15 downsamples the human representatives using the equal, mono, and para algorithms, respectively. This script is based on smot v0.14.2, the API may change in the future.
+Line 1 filters out all leaves with hosts other than swine or human. Line 2
+removes all monophyletic swine clades that have no representative from 2021.
+Line 3 removes all swine clades represented by a single member. Line 4 removes
+any coloring in the input tree. Lines 5 colors all remaining swine clades
+orange (hexadecimal code “#FFA000”). Line 6 colors leaves gray by default, then
+colors swine orange and finally recent swine blue. Lines 8-9, 10-11, and 14-15
+downsamples the human representatives using the equal, mono, and para
+algorithms, respectively. This script is based on smot v0.14.2, the API may
+change in the future.
 
 ![](images/pdm-0.png)
 
@@ -129,9 +137,8 @@ repo.
 
 ### `smot grep` - prune a tree to preserve only the taxa matching a pattern
 
-Create a subtree where all taxa are from the year 2020.The '-..-..' pattern is
-a regular expression that matches the MONTH-DAY syntax in a date and prevents
-most unwanted matches.
+Create a subtree where all taxa are from the year 2020.The `-..-..` pattern is
+a regular expression that matches the MONTH-DAY syntax in a date.
 
 ```
 smot grep --perl '2020-..-..' 1B.tre
@@ -144,7 +151,7 @@ and coloring. These factors can be specified in one of several ways.
 
  * `--factor-by-capture` uses a regular expression capture group to set the factor
  * `--factor-by-field` uses the index of a delimited field in the taxon label
- * `--factor-by-table` uses an two-column, TAB-delimited table that maps taxon name to factor
+ * `--factor-by-table` uses a two-column, TAB-delimited table that maps taxon name to factor
 
 The `smot factor` subcommand can list the extracted factors in a table:
 
@@ -163,8 +170,9 @@ taxon name (`smot factor append` and `smot factor prepend`, respectively).
 
 `smot` offers three general sampling algorithms.
 
- * The `equal` sampling method samples an equal number of taxa from each branch
-   descending from each node if sufficient taxa are available in each branch.
+ * The `equal` sampling method recursively samples an equal number of taxa from
+   each subtree descending from each node. If a subtree does contains too few
+   taxa, the remainder will be distributed to the sister subtrees.
 
  * The monophyletic sampling method (`mono`) subsamples a branch of the tree if
    all taxa within the branch have the same factor.
@@ -173,33 +181,34 @@ taxon name (`smot factor append` and `smot factor prepend`, respectively).
    monophyletic to be sampled together. The algorithm guarantees that at least
    one taxa is sampled from the deepest branch neighboring a different group.
 
-
-The following three commands all do the same thing (25% sampling from each of the 1B flu clades) using different factoring methods:
+The following three commands use the `mono` algorithm to sample 25% of the taxa
[…]
