### 2022-10-22 #4193 @Bisaloo
Thanks for the changes @arendsee, I believe this is a huge improvement and it will surely convince more potential users to give your tool a go.

@marekborowiec @Chjulian, could you have another go at reviewing this submission and update your respective checklists accordingly please? Thanks a lot for your help!

📝 [Checklist for Chjulian](https://github.com/openjournals/joss-reviews/issues/4193#issuecomment-1048945444)
📝 [Checklist for marekborowiec](https://github.com/openjournals/joss-reviews/issues/4193#issuecomment-1092189926)
