### 2022-11-25 #4193 @Chjulian
Dear all,

I tested smot on a macOS M1 with Monterey 12.6
I installed it without any warnings/errors.
I ran and checked all the examples provided, and some are not working as intended (see below)

--------
Errors:
I used the pdm.tre (ntax=247) stored in smot/test-data/ and the outputs have 233, two, three and four tips. Those outputs do not match figure (A). I suspect this is the wrong tree, given the number of taxa.

There is no explanation behind Example 2. The first two lines from example 2 return the following errors respectively:
Error: No such option: --seed Did you mean --keep?
Error: No such command 'prop'.

smot sample mono --seed=24601 --factor-by-field=6 -p 0.25 1B.tre
Returns the error:
IndexError: list index out of range
IndexError: Cannot access the 6th field in tip label 1.000
--------

The nested documentation is now evident in my head. The updated readme is more useful. However, as a returning user, I still found it challenging to understand what some of the commands do. Users would have to run each example command and check the trees in another program to figure out what changed in the trees. Also, the changes are not always readily evident because the example trees have many taxa (IMO, trees with tons of taxa are good for evaluating performance but not for showing how to use the program). So, small examples could be added in the future to increase visibility.
