### 2022-11-26 #4193 @arendsee
Thanks @Chjulian, I updated example 2. The example with `smot sample mono` shouldn't raise an error in the current version of `smot`. Previously `smot` died on failure to access a field by index in a taxa name. But that was a bit drastic, so now it will skip these cases.
