### 2022-12-06 0686524
Use \scriptsize for code to prevent overflow in pdf
