### 2022-12-14 #4193 @marekborowiec
@Bisaloo @arendsee Apologies for the long silence. I was able to install `smot` now and I think that documentation examples are much better now. Smaller trees and even more bite-sized examples would be even better but I guess what is there now is sufficient to use this program. I haven't been able to find the time to play around as extensively as before, but I noticed some deficiencies in the `smot color - color taxa labels or branches` README section: the first two examples are not in code blocks. Also, the 1B.tre from the example already appears to be colored this way so running the code doesn't change anything. Otherwise I think it looks good!
