### 2022-12-17 #4193 @arendsee
Thanks @Bisaloo, I've fixed the code blocks in the README. 1B.tre is a newick file without color in the current github version. Are you sure you didn't overwrite it locally?

You've put a lot of work into reviewing this package, I really appreciate it!
