### 2022-12-17 ddf49f8
Release v1.0.0 - JOSS publication
