+++ ChangeLog.md
@@ -1,3 +1,8 @@
+1.0.0 [2022-12-17]
+===================
+
+JOSS publication version
+
 
 0.17.5 [2022-05-29]
 ===================
