+++ paper/paper.bib
@@ -25,7 +25,7 @@
 }
 
 @article{anderson2016phylogeny,
-  title={A phylogeny-based global nomenclature system and automated annotation tool for H1 hemagglutinin genes from swine influenza A viruses},
+  title={A phylogeny-based global nomenclature system and automated annotation tool for H1 hemagglutinin genes from swine influenza {A} viruses},
   author={Anderson, Tavis K and Macken, Catherine A and Lewis, Nicola S and Scheuermann, Richard H and Van Reeth, Kristien and Brown, Ian H and Swenson, Sabrina L and Simon, Ga{\"e}lle and Saito, Takehiko and Berhane, Yohannes and others},
   journal={MSphere},
   volume={1},
@@ -51,7 +51,7 @@
 }
 
 @article{vijaykrishna2010reassortment,
-  title={Reassortment of pandemic H1N1/2009 influenza A virus in swine},
+  title={Reassortment of pandemic H1N1/2009 influenza {A} virus in swine},
   author={Vijaykrishna, D and Poon, LLM and Zhu, HC and Ma, SK and Li, OTW and Cheung, CL and Smith, GJD and Peiris, JSM and Guan, Y},
   journal={Science},
   volume={328},
@@ -64,7 +64,7 @@
 }
 
 @article{smith2009origins,
-  title={Origins and evolutionary genomics of the 2009 swine-origin H1N1 influenza A epidemic},
+  title={Origins and evolutionary genomics of the 2009 swine-origin H1N1 influenza {A} epidemic},
   author={Smith, Gavin JD and Vijaykrishna, Dhanasekaran and Bahl, Justin and Lycett, Samantha J and Worobey, Michael and Pybus, Oliver G and Ma, Siu Kit and Cheung, Chung Lam and Raghwani, Jayna and Bhatt, Samir and others},
   journal={Nature},
   volume={459},
@@ -76,7 +76,7 @@
 }
 
 @article{nelson2015continual,
-  title={Continual reintroduction of human pandemic H1N1 influenza A viruses into swine in the United States, 2009 to 2014},
+  title={Continual reintroduction of human pandemic H1N1 influenza {A} viruses into swine in the {U}nited {S}tates, 2009 to 2014},
   author={Nelson, Martha I and Stratton, Jered and Killian, Mary Lea and Janas-Martindale, Alicia and Vincent, Amy L},
   journal={Journal of virology},
   volume={89},
@@ -202,7 +202,7 @@
 }
 
 @article{sukumaran2010dendropy,
-  title={DendroPy: a Python library for phylogenetic computing},
+  title={DendroPy: a {P}ython library for phylogenetic computing},
   author={Sukumaran, Jeet and Holder, Mark T},
   journal={Bioinformatics},
   volume={26},
@@ -214,7 +214,7 @@
 }
 
 @article{cock2009biopython,
-  title={Biopython: freely available Python tools for computational molecular biology and bioinformatics},
+  title={Biopython: freely available {P}ython tools for computational molecular biology and bioinformatics},
   author={Cock, Peter JA and Antao, Tiago and Chang, Jeffrey T and Chapman, Brad A and Cox, Cymon J and Dalke, Andrew and Friedberg, Iddo and Hamelryck, Thomas and Kauff, Frank and Wilczynski, Bartek and others},
   journal={Bioinformatics},
   volume={25},
@@ -250,7 +250,7 @@
 
 @article{han2019phyclip,
     author = {Han, Alvin X and Parker, Edyth and Scholer, Frits and Maurer-Stroh, Sebastian and Russell, Colin A},
-    title = "{Phylogenetic Clustering by Linear Integer Programming (PhyCLIP)}",
[…]
