### 2022-12-20 #4193 @danielskatz
@arendsee - the track editor for this paper is currently out, so I'm helping. I've proofread your paper and have a few small changes in https://github.com/flu-crew/smot/pull/7 - please merge this or let me know you disagree with.  In addition, there's a word missing at the end of the first sentence of the caption of Figure 3.  Please let me know when this is fixed.
