### 2022-12-20 #4193 @danielskatz
@openjournals/dev - Can you see what the issue is with the PDF, which I don't see, though the DOI resolves correctly?
