### 2022-12-20 #4193 @danielskatz
There was a problem with GitHub pages that's been fixed, so this should now be working. It still isn't working for me, but it is for other people, so it appears I have a caching issue. So I'm going to go ahead and close the issue.
