### 2022-12-20 #4193 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
